requirements.txt
pandas>=2.0.0
numpy>=1.24.0
faker>=18.0.0
pyyaml>=6.0
pytest>=7.3.0
pyarrow>=12.0.0
